Move BindingReflectionHintsRegistrar to spring-context
Closes gh-28735
